msp/alertpolicy: unify on generalized policy builder, with custom condition builders
