+++ dev/managedservicesplatform/internal/resource/alertpolicy/BUILD.bazel
@@ -3,7 +3,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "alertpolicy",
-    srcs = ["alertpolicy.go"],
+    srcs = [
+        "alertpolicy.go",
+        "responsecode.go",
+        "thresholdaggregation.go",
+    ],
     importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/resource/alertpolicy",
     visibility = ["//dev/managedservicesplatform:__subpackages__"],
     deps = [
@@ -19,7 +23,10 @@ go_library(
 
 go_test(
     name = "alertpolicy_test",
-    srcs = ["alertpolicy_test.go"],
+    srcs = [
+        "responsecode_test.go",
+        "thresholdaggregation_test.go",
+    ],
     embed = [":alertpolicy"],
     deps = [
         "//lib/pointers",
